fix: resolve OOM and compatibility issues in CPT pipeline
- Reduce per_device_train_batch_size 8→2, gradient_accumulation 4→16 (OOM fix)
- Add PYTORCH_CUDA_ALLOC_CONF=expandable_segments:True
- Copy data/ into Docker image so training data is available at runtime
- Fix torch_dtype→dtype deprecation in train.py
- Pin transformers>=4.51.0, add tiktoken (required by Qwen3.5 tokenizer)
- Install torch==2.5.1 via cu121 wheel separately (avoids version conflicts)

+++ cpt/train.py
@@ -79,7 +79,7 @@ def main() -> None:
 
     print(f"Loading model: {model_cfg['name']}")
     dtype = torch.bfloat16 if train_cfg["bf16"] else torch.float32
-    model = AutoModelForCausalLM.from_pretrained(model_cfg["name"], torch_dtype=dtype)
+    model = AutoModelForCausalLM.from_pretrained(model_cfg["name"], dtype=dtype)
     if train_cfg.get("gradient_checkpointing"):
         model.gradient_checkpointing_enable()
